pyproject.toml
[project]
name = "legal-benchmark-runner"
version = "0.1.0"
description = "Minimal multi-provenance legal benchmark runner"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
  "boto3>=1.34.0",
  "google-genai>=1.0.0",
  "litellm>=1.0.0",
  "pypdf>=4.0.0",
  "PyYAML>=6.0",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.uv]
package = false
